gitignore

+++ huis.sh
@@ -1,5 +1,9 @@
 #!/usr/bin/env bash
 
-for symlink in .tmux.conf .vimrc .spacemacs ; do
-  ln -s $HOME/.huis/modules/huis-dotfiles/$symlink $HOME/$symlink
+for symlink in .tmux.conf .vimrc .spacemacs .gitignore; do
+  if [ ! -L $HOME/$symlink ] ; then
+    ln -s $HOME/.huis/modules/huis-dotfiles/$symlink $HOME/$symlink
+  fi
 done
+
+git config --global core.excludesfile '~/.gitignore'
